+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - Fixed bug where color automation did not work on Linux because of line ending character.
+- Fix: Control + Up/Down don't works as expected on neovim [#371](https://github.com/raphamorim/rio/issues/371) 
+- Fix: remove duplicate kitty backspace keybinds [#375](https://github.com/raphamorim/rio/pull/375) by [@hougesen](https://github.com/hougesen)
+- Fix: Kitty-keyboard-protocol causes Backspace to delete 2 characters. [#344](https://github.com/raphamorim/rio/issues/344) by [@hougesen](https://github.com/hougesen)
 
 ## 0.0.30